+++ musicdl/modules/common/__init__.py
@@ -1,2 +1,3 @@
 '''initialize'''
+from .tunehub import TuneHubMusicClient
 from .gdstudio import GDStudioMusicClient